Move config-related stuff out of main.rs
